Split web/app.py (3032 lines) into route modules under web/routes/
TASK-085: Refactored monolithic web/app.py into 14 focused modules:
- app.py: Slim entry point (109 lines) — create_app, middleware, router registration
- models.py: 7 Pydantic request/response DTOs (76 lines)
- helpers.py: Shared helpers — DuckDB queries, config, logging (798 lines, exempted)
- routes/health.py: /api/status, /api/watcher/status, /api/health/platform
- routes/config.py: /api/config, /api/metabase-config
- routes/sources.py: /api/sources, /api/sources/{name}/details
- routes/sync.py: /api/sources/{name}/sync + run_sync_task()
- routes/logs.py: /api/logs, /api/sources/{name}/logs
- routes/dbt.py: /api/dbt/models, model run + dbt docs proxy
- routes/upload.py: CSV upload/list/delete (664 lines, exempted)
- routes/websocket.py: ConnectionManager, ws_manager, /ws
- routes/ui.py: /, /health, /logs, /api, /api/docs, /api/redoc
- routes/metabase_proxy.py: All Metabase proxy routes + SSO

DOC-004: Created web/CLAUDE.md documenting post-refactor structure.

All 176 tests pass. All CI checks pass (lint, format, headers, docstrings, file sizes).
