+++ dango/web/routes/__init__.py
@@ -1,0 +1,4 @@
+"""dango/web/routes/__init__.py
+
+Route modules for the Dango web API.
+"""
