+++ docs/file-exemptions.yml
@@ -10,13 +10,18 @@ schema_version: 1
 limit: 500  # Informational — scripts/check_file_sizes.py owns the enforced threshold
 
 exemptions:
-  # --- Refactoring planned in v1 ---
-  - file: dango/web/app.py
-    lines: 2900
-    category: refactor
-    task: TASK-085
-    reason: "Monolithic web server; will be split into web/routes/ submodules"
-    review_by: "v1.0"
+  # --- Post-TASK-085 split results (from web/app.py 3032 → web/routes/) ---
+  - file: dango/web/helpers.py
+    lines: 798
+    category: exempt
+    reason: "Shared helper functions for web routes. Consolidates DuckDB queries, config loading, service health checks, and log management. Extracted from web/app.py by TASK-085. Mirrors cli/utils.py pattern."
+    review_by: "v1.1"
+
+  - file: dango/web/routes/upload.py
+    lines: 664
+    category: exempt
+    reason: "CSV upload/list/delete endpoints. Extracted from web/app.py by TASK-085. DELETE endpoint alone has ~300 lines of DuckDB cleanup logic."
+    review_by: "v1.1"
 
   # --- Post-TASK-005 split results (from cli/main.py 4119 → cli/commands/) ---
   - file: dango/cli/commands/platform.py
